Rename "blanket" to "bridge"
This name is more specific and therefore descriptive. I think it fits
because these impls "bridge" from Box<impl Trait> to impl Trait. I also
changed to using parentheses, because it allows us to extend the grammar
in more ways in the future (e.g. see the last step of #95), and it
avoids muddying the meaning of = in the base syntax.